UC14 Toevoeging Test + Verandering datatype
Product Price property datatype veranderd van double naar decimal.

Test toegevoegd om formatting te checken.

+++ Grocery.Core/Models/Product.cs
@@ -7,11 +7,11 @@ namespace Grocery.Core.Models
         [ObservableProperty]
         public int stock;
 
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public DateOnly ShelfLife { get; set; }
-        public Product(int id, string name, int stock, double price) : this(id, name, stock, default, price) { }
+        public Product(int id, string name, int stock, decimal price) : this(id, name, stock, default, price) { }
 
-        public Product(int id, string name, int stock, DateOnly shelfLife, double price) : base(id, name) 
+        public Product(int id, string name, int stock, DateOnly shelfLife, decimal price) : base(id, name) 
         {
             Stock = stock;
             ShelfLife = shelfLife;
